services/api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg==3.2.3
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.9
openpyxl==3.1.5
pytest==8.3.3
httpx==0.27.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4